fix implicit conversion warnings

+++ Source/PluginEditor.h
@@ -144,8 +144,8 @@ struct AnalyzerPathGenerator
             {
                 auto binFreq = binNum * binWidth;
                 auto normalizedBinX = juce::mapFromLog10(binFreq, 20.f, 20000.f);
-                int binX = std::floor(normalizedBinX * width);
-                p.lineTo(binX, y);
+                int binX = std::floor(int(normalizedBinX * width));
+                p.lineTo(float(binX), float(y));
             }
         }
